+++ NEWS.md
@@ -22,6 +22,7 @@
 * `geom_sina()` now works when the group only have a single unique y value
 * `facet_zoom()` now works correctly with transformed scales
 * `facet_wrap_paginate()` now works correctly with `dir = 'v'`
+* `facet_matrix()` now supports a labeller
 
 # ggforce 0.3.4